refactor(emitter): split fix verification generic call tests
